Move LateJoinGui to the UI folder
